builder/requirements.txt
genanki>=0.13
pandas>=2.0
gTTS>=2.4
Pillow>=10.0
pyyaml>=6.0
elevenlabs>=1.0.0
